# making marker dropbomb optional on release