We added https://github.com/matklad/drop_bomb to ```rslint_parser::Marker``` to, very politely, panic if you forget to deal correctly with ```Marker```s. 

It is really helpful when developing but it has a performance impact that we can avoid in release mode. This PR makes the bomp do nothing in release, letting the optimizer to completely scrap it, but also allow the bomb to be opted-in by the ```force-markerbomb``` feature.

The performance gain is smaller than when I investigated this for the first time, but still nice to have.

Before
```
https://cdn.jsdelivr.net/npm/three@0.134.0/build/three.min.js
                        time:   [84.660 ms 86.023 ms 87.527 ms]
```

After
```
https://cdn.jsdelivr.net/npm/three@0.134.0/build/three.min.js
                        time:   [82.035 ms 82.676 ms 83.358 ms]
```

The bomb can be forced using

```
cargo run -p xtask --features "rslint_parser/force-markerbomb" --release -- coverage-libs --filter=three
```

for more see: https://github.com/rome/tools/issues/1901