@Boshen 2022-01-07
Splitting out a "debug release" profile with custom overrides will make things easier, i.e. https://doc.rust-lang.org/cargo/reference/profiles.html#overrides