@xunilrj 2022-01-07
> I believe the crate has already built in support for disabling drop bombs with a feature flag

You are right. Still don't know why I saw better performance when I commented these bombs. Will close this.